refactor to cache, move to commands
